zprmonitor
-usuniecie kilku warningow z windowsa

+++ code/src/CpuDaemon.cpp
@@ -64,7 +64,7 @@ void CpuDaemon::doMeasure() {
     init();
     if (cpuQuery != 0) {
         std::this_thread::sleep_for(std::chrono::milliseconds(50));
-        actValue_ = getCurrentValue();
+        actValue_ = static_cast<int>(getCurrentValue());
         signal_(actValue_); 
     }
     else {
